Rename Either to Auto
