+++ src/types/path/auto.ts
