+++ src/types/path/index.ts
@@ -3,12 +3,12 @@ import { Primitive } from '../utils';
 
 import { TupleKeys } from './internal/keys';
 import { ArrayKey, IsTuple } from './internal/utils';
+import type * as Auto from './auto';
 import type * as Branded from './branded';
-import type * as Either from './either';
 import type * as Lazy from './lazy';
 
 /** Re-export public API */
-export type { Branded, Either, Lazy };
+export type { Auto, Branded, Lazy };
 export type { PathString } from './pathString';
 
 /**
